Orders list

+++ app/Http/Livewire/ProductsList.php
@@ -72,12 +72,26 @@ class ProductsList extends Component
     {
         $product = Product::findOrFail($id);
 
+        if ($product->orders()->exists()) {
+            $this->addError('orderexist', 'This product cannot be deleted, it already has orders');
+            return;
+        }
+
         $product->delete();
     }
 
     public function deleteSelected(): void
     {
-        $products = Product::whereIn('id', $this->selected)->get();
+        $products = Product::with('orders')->whereIn('id', $this->selected)->get();
+
+        foreach ($products as $product) {
+            if ($product->orders()->exists()) {
+                $this->addError("orderexist", "Product <span class='font-bold'>{$product->name}</span> cannot be deleted, it already has orders");
+                return;
+            }
+        }
+
+        $products->each->delete();
 
         $this->reset('selected');
     }
